Use temporary github dependencies
Of course I am using linked packages during development, in our case
here with domainql-form and automaton-js to get a three watch-job
environment going where I can just edit any source file and webpack
will update.

This meant however, that the actual dependencies in the package.json
remained frozen, which I recognize now is not the best way to do it
in a collaborative project.

So we're moving to direct dependencies from github for now. We're hosted
there anyway and this way, the package.json expresses as working setup
sans links as long as all branches are up-to-date.

And we can directly link them anyhow.