Implement LIS-A validation prototype
Implements logic to provide a rules.json file and to hot-load it both as
external file and as servlet resource.

Moves the model classes to a proper model package.

Adds the newly provided `validationRules` as global config.

Adds a rules.json overlay for our FooForm and configures the form to
look it up by name.