Implement ScrollTracker component
This commits adds a somewhat experimental feature of a scroll tracker
component that prevents changes in form errors to cause the currently
focused input to jump on the screen.

To minimize distraction, this snapshots to current screen position of
the currently focused input before an update and restores it afterwards,
potentially by introducing additional space above or below the current
screen form to have additional scrolling space.

This uses a variant of domainql-form's withForm which is kind of
copy-and-paste as I could not find a quick and acceptable form to reuse
withForm while introducing all the parts withAppForm needs.

Might improve later.