Integrate schema docs
Use the new TypeDoc extraction feature to improve the documentation of
the test project.

We use a combination of multiple type doc sources for the project

 * The typeDocs as extracted from the local sources
 * Hand-written type documentation to augment the database types
 * Typedoc data included in the automaton project

For the local source docs, we add a exec:java execution to our maven
POM.

On the "generate-resources" phase, the plugin extracts the Javadoc
information of our model and service folders into the TypeDoc JSON
format. This JSON file is deployed with the (exploded) WAR and is used
at runtime to augment the autogenerated schema.