Remove heading from Tree
There seemed to be some odd problem with the heading acting as a tab
catcher for some reason.

We don't really need a heading and had only included it to mirror the
original WAI-ARIA best practice example.

So I'm just removing the heading for now.

If the tester prefers there to be a heading, we could also try just
removing the "sr-only" class.