Implement process v2 transformer
Implements a jscodeshift transformer that does most of the work to
convert the old process format into the new process format.

It splits up the process modules into separate state modules and merges
those with the composite model for it.

It removes the composite resource files only leaving outdated <XXXForm/>
components behind which should be either dissolved or moved.

The startup function variant is removed and the code for that is moved
directly into the initProcess() function of the process.

It transforms "t.context" and "t.target" references into local "context"
and "target" variables and return the target variable at the end.

It fixes all imports when we move code to the states directory.

It replaces all exact string literal matches of state names with an
imported state name literal.

Usage
=====

Make sure you have a clean working directory with all changes commited
to be sure.

From the root of the automaton application project execute

jscodeshift -t mpath/process-v2-codemod.js src/main/js/apps/myapp/processes

replace `mpath` with the path to the codmod and `myapp` with the name
of your app(s)

Note that the transformation is not perfect. Some start functions cannot
be ported to the new system as-is. I hope I got all the imports right
now.