CommonNav: List processes dynamically
We always keep forgetting to add processes and this app is just a
collection of functionality tests anyway.

Uses the new ProcessListUtil to add a "processes" key to our initial
data and config.

Adds an ignore list for processes that require parametrization. Those
drop down items need to be added manually.