FK-Test: Add validation test view
Adds an extra view to fk-test to test the new enhanced validation
behavior for <FKSelector/>

Uses a local/compenent level validator and a local form context with a
high-level validation on the same field.

From the current db dump, the 'Qux A #1' entry triggers the local one
and the 'Qux A #2' the high-level validation.